refactor(account-data): own account data model types
