+++ src/services/apiAdapters/contracts/accountRefresh.ts
@@ -1,7 +1,5 @@
-import type {
-  ApiServiceAccountRequest,
-  RefreshAccountResult,
-} from "~/services/apiService/common/type"
+import type { RefreshAccountResult } from "~/services/accountData/model"
+import type { ApiServiceAccountRequest } from "~/services/apiService/common/type"
 import type { ApiServiceRequest } from "~/services/apiTransport/type"
 
 export type AccountRefreshSupportRequest = Pick<
