+++ src/services/apiService/anyrouter/index.ts
@@ -1,14 +1,14 @@
+import type {
+  AccountData,
+  RefreshAccountResult,
+} from "~/services/accountData/model"
 import {
   determineHealthStatus,
   fetchAccountQuota,
   fetchTodayIncome,
   fetchTodayUsage,
 } from "~/services/apiService/common"
-import type {
-  AccountData,
-  ApiServiceAccountRequest,
-  RefreshAccountResult,
-} from "~/services/apiService/common/type"
+import type { ApiServiceAccountRequest } from "~/services/apiService/common/type"
 import type { ApiServiceRequest } from "~/services/apiTransport/type"
 import { anyrouterProvider } from "~/services/checkin/autoCheckin/providers/anyrouter"
 import { SiteHealthStatus, type CheckInConfig } from "~/types"
