+++ src/services/apiService/doneHub/index.ts
@@ -1,3 +1,7 @@
+import type {
+  AccountData,
+  RefreshAccountResult,
+} from "~/services/accountData/model"
 import {
   determineHealthStatus,
   fetchAccountQuota,
@@ -8,11 +12,7 @@ import {
 import { REQUEST_CONFIG } from "~/services/apiService/common/constant"
 import { ApiError } from "~/services/apiService/common/errors"
 import { fetchAllItems } from "~/services/apiService/common/pagination"
-import type {
-  AccountData,
-  ApiServiceAccountRequest,
-  RefreshAccountResult,
-} from "~/services/apiService/common/type"
+import type { ApiServiceAccountRequest } from "~/services/apiService/common/type"
 import { fetchApi, fetchApiData } from "~/services/apiService/common/utils"
 import type { ApiServiceRequest } from "~/services/apiTransport/type"
 import type { TodayLogQueryConfig } from "~/services/usageLogs/model"
